修改废弃的 v8 函数。

+++ fibjs/src/DBRow.cpp
@@ -36,7 +36,7 @@ result_t DBRow::_named_enumerator(v8::Handle<v8::Array>& retVal)
 	int32_t i;
 
 	retVal = v8::Array::New();
-	for (i = 0; i < m_cols.size(); i++)
+	for (i = 0; i < (int)m_cols.size(); i++)
 		retVal->Set(i, m_cols[i]);
 
 	return 0;
